+++ src/ipc/utils/git_utils.ts
@@ -8,6 +8,7 @@ import pathModule from "node:path";
 import { readSettings } from "../../main/settings";
 import log from "electron-log";
 import { normalizePath } from "../../../shared/normalizePath";
+import type { UncommittedFile, UncommittedFileStatus } from "../ipc_types";
 const logger = log.scope("git_utils");
 import type {
   GitBaseParams,
@@ -416,16 +417,8 @@ export async function getGitUncommittedFiles({
   }
 }
 
-export type UncommittedFileStatus =
-  | "added"
-  | "modified"
-  | "deleted"
-  | "renamed";
-
-export interface UncommittedFile {
-  path: string;
-  status: UncommittedFileStatus;
-}
+// Re-export from ipc_types for backwards compatibility
+export type { UncommittedFile, UncommittedFileStatus } from "../ipc_types";
 
 /**
  * Get uncommitted files with their status (added, modified, deleted, renamed).
